+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+* Added `quote::ToTokens` bound to `MaybeEnum`.
+
+* Improved error messages.
+
 # 0.7.2 - 2019-05-30
 
 * Improved error messages.